+++ modules/core/shared/src/test/scala/gem/arb/ArbStep.scala
@@ -4,9 +4,8 @@
 package gem
 package arb
 
-import gem.enum.SmartGcalType
-import gem.Step
-import gem.config.{ DynamicConfig, GcalConfig, TelescopeConfig }
+import gem.enum.{Instrument, SmartGcalType}
+import gem.config.{GcalConfig, TelescopeConfig}
 import org.scalacheck._
 import org.scalacheck.Arbitrary._
 
@@ -17,22 +16,31 @@ trait ArbStep extends Arbitraries {
   import ArbDynamicConfig._
   import ArbTelescopeConfig._
 
-  val genBase: Gen[Step.Base] =
-    Gen.oneOf(
-      Gen.const(Step.Base.Bias),
-      Gen.const(Step.Base.Dark),
-      arbitrary[GcalConfig].map(Step.Base.Gcal(_)),
-      arbitrary[TelescopeConfig].map(Step.Base.Science(_)),
-      arbitrary[SmartGcalType].map(Step.Base.SmartGcal)
-    )
+  implicit val arbStepBase: Arbitrary[Step.Base] =
+    Arbitrary {
+      Gen.oneOf(
+        Gen.const(Step.Base.Bias),
+        Gen.const(Step.Base.Dark),
+        arbitrary[GcalConfig].map(Step.Base.Gcal(_)),
+        arbitrary[TelescopeConfig].map(Step.Base.Science(_)),
+        arbitrary[SmartGcalType].map(Step.Base.SmartGcal)
+      )
+    }
+
+  def genStepOf(i: Instrument): Gen[Step] =
+    for {
+      b <- arbitrary[Step.Base]
+      d <- genDynamicConfigOf(i)
+    } yield d.toStep(b)
 
   implicit val arbStep: Arbitrary[Step] =
-    Arbitrary {
-      for {
-        dc <- arbitrary[DynamicConfig]
-        b  <- genBase
-      } yield dc.toStep(b)
-    }
+    Arbitrary(arbitrary[Instrument].flatMap(genStepOf))
+
+  def genSequenceOf(i: Instrument): Gen[List[Step]] =
+    for {
+      n <- Gen.choose(0, 50)
+      s <- Gen.listOfN(n, genStepOf(i))
+    } yield s
 
 }
 object ArbStep extends ArbStep
